feat(ui): update styles + try dynamic yaml import
