--- ui/src/layout/Scans/ScanConfigWizardModal/components/CustomQueryBuilder.functions.js
@@ -1,21 +1,0 @@
-const convertDataType = (originalType) => {
-    switch (originalType) {
-        case "string":
-            return "text";
-        case "integer":
-            return "number";
-        case "object":
-            return "!struct";
-        case "array":
-            return "!struct"
-        default:
-            return originalType;
-    }
-}
-
-const postConvertQuery = (query) => {
-    let result = query?.replace(/&&/g, "and");
-    return result;
-}
-
-export { convertDataType, postConvertQuery };
