+++ ui/src/layout/Scans/ScanConfigWizardModal/scan-config-wizard-modal.scss
@@ -7,6 +7,8 @@
         }
     }
     .scan-config-general-step {
+        flex: 1 1 auto;
+        
         .scan-config-instances .form-field-wrapper {
             margin-bottom: 17px;
         }
